+++ pytket/tests/circuit_test.py
@@ -740,6 +740,28 @@ def test_clifford_checking() -> None:
     assert m.is_clifford_type() == False
 
 
+def test_getting_registers() -> None:
+    c = Circuit(2, 1)
+    q_err_msg = "Cannot find quantum register with name"
+    c_err_msg = "Cannot find classical register with name"
+    with pytest.raises(RuntimeError) as e:
+        c.get_c_register("q")
+    assert c_err_msg in str(e.value)
+    with pytest.raises(RuntimeError) as e:
+        c.get_q_register("c")
+    assert q_err_msg in str(e.value)
+    assert c.get_c_register("c").name == "c"
+    assert c.get_c_register("c").size == 1
+    assert c.get_q_register("q").name == "q"
+    assert c.get_q_register("q").size == 2
+    c.add_q_register("test_qr", 10)
+    c.add_c_register("test_cr", 8)
+    assert c.get_c_register("test_cr").name == "test_cr"
+    assert c.get_c_register("test_cr").size == 8
+    assert c.get_q_register("test_qr").name == "test_qr"
+    assert c.get_q_register("test_qr").size == 10
+
+
 if __name__ == "__main__":
     test_circuit_gen()
     test_symbolic_ops()